Add reply menu
This commit adds a fully functional reply menu
that will only show up if the user has a youtube link
in their reply. It lets the user decide what time
he/she wants the link to start at,
and even lets them decide whether or not they
want the link extracted and placed in a button

Graphics for desktop and mobile supported.